Working tilemap

+++ Timewarp Engine/Timewarp Engine/include/TIMEWARP ENGINE/loadTilemap.h
@@ -1,4 +1,4 @@
 #pragma once
 #include <vector>
 
-std::vector<std::vector<int>> loadLevel(int levelID);
+std::vector<std::vector<float>> loadLevel(int levelID);
